Serialization fix

+++ NAMSC_lib/src/Novel/Event/Event.cpp
@@ -17,8 +17,8 @@ void swap(Event& first, Event& second) noexcept
 }
 
 Event::Event(Scene* const parentScene, const QString& label, std::vector<std::unique_ptr<Action>>&& actions)
-	: parentScene(parentScene), 
-	label(label), 
+	: parentScene(parentScene),
+	label(label),
 	actions_(std::move(actions))
 {
 }
@@ -27,9 +27,9 @@ void Event::serializableLoad(QDataStream& dataStream)
 {
 	dataStream >> label;
 
-	uint actionsSize;
+	size_t actionsSize;
 	dataStream >> actionsSize;
-	for (uint i = 0u; i != actionsSize; ++i)
+	for (size_t i = 0u; i != actionsSize; ++i)
 	{
 		NovelLib::SerializationID type;
 		dataStream >> type;
